Fix #8: add a pager simulation.
This might need to be extended to support the concept of mandatory snap points, as the CSS snap scroll proposal allows.

I also tweaked the implementation of Spring to follow clippy's recommendation on float comparisons. This change also adds some imports from the standard library. It's using the core versions, so we could do a no_std version as well.